Step 14, refactoring, loading students file by default, prompting file name, and using Ruby's CSV Class

+++ directory.rb
@@ -1,22 +1,28 @@
+require 'csv'
+
 @students = []
+@cohort = :november
 
-def input_students
-  puts "Please enter the names of the students"
-  puts "To finish, just hit return twice"
-  name = STDIN.gets.chomp
-  
-  while !name.empty? do
-    @students << {name: name, cohort: :november}
-    if @students.length > 1
-      puts "Now we have #{@students.count} students"
-    else
-      puts "Now we have #{@students.count} student"
-    end
-
-    name = STDIN.gets.chomp
+def load_students_on_open
+  filename = "students.csv"
+  if File.exist?(filename)
+    load_students(filename)
+      puts "Loaded #{@students.count} from #{filename}"
+  else
+    return if filename.nil?
   end
+end
 
-end 
+def add_student(name, cohort)
+  @students << {name: name, cohort: cohort.to_sym}
+end
+
+def load_students(filename)
+  CSV.foreach(filename) { |row|
+    name, cohort = row
+    add_student(name, cohort)
+  }
+end
 
 def interactive_menu
   loop do
@@ -25,42 +31,34 @@ def interactive_menu
   end
 end
 
-def try_load_students
-  filename = ARGV.first
-  return if filename.nil?
-  if File.exist?(filename)
-    load_students(filename)
-      puts "Loaded #{@students.count} from #{filename}"
-  else
-    puts "Sorry, #{filename} doesn't exist."
-    exit
-  end
-end
-
 def print_menu
   puts "1. Input the students"
   puts "2. Show the students"
-  puts "3. Save the list to students.csv"
-  puts "4. Load the list from students.csv"
+  puts "3. Save the list"
+  puts "4. Load the list"
   puts "9. Exit" 
-end
-
-def show_students
-  print_header
-  print_student_list
-  print_footer
 end
 
 def process(selection)
   case selection
   when "1"
     input_students
+    puts "Input successful!"
   when "2"
     show_students
+    puts "Showing students successful!"
   when "3"
     save_students
+    puts "Save successful!"
   when "4"
-    load_students
+    puts "Enter the name and extension of the file you'd like to load"
+    file = gets.chomp
+    if File.exist?(file)
+      load_students(file)
+      puts "Load successful!"
+    else
+      puts "Sorry, #{file} doesn't exist"
+    end
   when "9"
     exit 
   else
@@ -68,9 +66,31 @@ def process(selection)
   end
 end
 
-def print_header
+def input_students
+  puts "Please enter the names of the students"
+  puts "To finish, just hit return twice"
+  name = STDIN.gets.chomp
+  
+  while !name.empty? do
+    add_student(name, @cohort)
+    if @students.count > 1
+      puts "Now we have #{@students.count} students"
+    else
+      puts "Now we have #{@students.count} student"
+    end
+
+    name = STDIN.gets.chomp
+  end
+
+end
+
+def show_students
   puts "The students of Villains Academy"
   puts "-------------"
+  
+  print_student_list
+  
+  puts "Overall, we have #{@students.first.count} great students"
 end
 
 def print_student_list
@@ -79,30 +99,24 @@ def print_student_list
   }
 end
 
-def print_footer
-  puts "Overall, we have #{@students.count} great students"
-end
-
 def save_students
-  # open the file for writing
-  file = File.open("students.csv", "w")
-  # iterate over the aray of students
-  @students.each { |student|
-    student_data = [student[:name], student[:cohort]]
-    csv_line = student_data.join(",")
-    file.puts csv_line
+  puts "Please enter the name and extension of the file you'd like to save as"
+  file_name = gets.chomp
+  if File.exist?(file_name)
+    CSV.open("#{file_name}", "w") { |row|
+      @students.each { |student|
+        row << [student[:name], student[:cohort]]
+      }
+    }
+  else
+    file = File.open("students.csv", "w")
+    @students.each { |student|
+      student_data = [student[:name], student[:cohort]]
+      csv_line = student_data.join(",")
+      file.puts csv_line
   }
-  file.close
+  end
 end
 
-def load_students(filename = "students.csv")
-  file = File.open(filename, "r")
-  file.readlines.each { |line|
-    name, cohort = line.chomp.split(',')
-    @students << {name: name, cohort: cohort.to_sym}
-  }
-  file.close
-end
-
-try_load_students
+load_students_on_open
 interactive_menu
